refactor(react-server): organize files by features
