+++ packages/react-server/src/entry/browser.tsx
@@ -3,9 +3,9 @@ import { createBrowserHistory } from "@tanstack/history";
 import React from "react";
 import reactDomClient from "react-dom/client";
 import { rscStream } from "rsc-html-stream/client";
+import { initializeWebpackBrowser } from "../features/use-client/browser";
 import { RootErrorBoundary } from "../lib/client/error-boundary";
 import { Router, RouterContext, useRouter } from "../lib/client/router";
-import { initDomWebpackCsr } from "../lib/csr";
 import { __global } from "../lib/global";
 import { injectActionId, wrapRscRequestUrl } from "../lib/shared";
 import type { CallServerCallback } from "../lib/types";
@@ -17,7 +17,7 @@ export async function start() {
     return;
   }
 
-  initDomWebpackCsr();
+  initializeWebpackBrowser();
 
   const { default: reactServerDomClient } = await import(
     "react-server-dom-webpack/client.browser"
