+++ packages/react-server/src/features/README.md
@@ -1,0 +1,2 @@
+Rakkasjs-style file organization.
+see https://github.com/rakkasjs/rakkasjs/tree/main/packages/rakkasjs/src/features
