+++ packages/react-server/src/features/server-action/browser.tsx
@@ -1,0 +1,1 @@
+// TODO: callServer etc...
